+++ cmake/BipedalLocomotionFrameworkDependencies.cmake
@@ -168,6 +168,10 @@ framework_dependent_option(FRAMEWORK_COMPILE_FloatingBaseEstimators
   "Compile FloatingBaseEstimators libraries?" ON
   "FRAMEWORK_USE_manif;FRAMEWORK_COMPILE_Contact" OFF)
 
+framework_dependent_option(FRAMEWORK_COMPILE_RobotDynamicsEstimator
+  "Compile RobotDynamicsEstimator libraries?" ON
+  "" OFF)
+
 framework_dependent_option(FRAMEWORK_COMPILE_ManifConversions
   "Compile manif Conversions libraries?" ON
   "FRAMEWORK_USE_manif" OFF)